requirements.txt
aiofile==1.5.1
aiofiles==0.4.0
aiohttp==3.5.0
aiohttp-socks==0.2.1
aiozeroconf==0.1.8
altgraph==0.16.1
ansicolors==1.1.8
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
black==18.9b0
certifi==2018.10.15
cffi==1.7.0
chardet==3.0.4
Click==7.0
crypto==1.4.1
cryptography==2.4.2
filetype==1.0.1
future==0.17.1
humanize==0.5.1
idna==2.7
ifaddr==0.1.4
ipaddress==1.0.22
macholib==1.11
multidict==4.4.2
Naked==0.1.31
names==0.3.0
netifaces==0.10.7
ordered-set==3.0.2
pefile==2018.8.8
prompt-toolkit==2.0.5
pycparser==2.19
PyNaCl==1.3.0
python-libmagic==0.4.0
PyYAML==3.13
requests==2.20.0
shellescape==3.4.1
six==1.11.0
stem==1.6.0
toml==0.10.0
tqdm==4.28.1
typed-bencode==0.0.1
urllib3==1.24
wcwidth==0.1.7
yarl==1.2.6
